Second Minimum Node In a Binary Tree - Accepted Solution
Trick : the root is the smallest value present in the tree. Second smallest will be found either in left or right subtree. Also, as soon as different node is encountered, no need to go downwards as the it is confirmed that the value is only going to increase (no scope of finding second SMALLEST value there!) 
In short, 2nd smallest value will occur at the 1st point of difference in BOTH the subtrees. 
Time Complexity : O(x) where x is the number of elements not equal to root.val
Space Complexity : O(h) where h is height of the tree used up by stack of recursion